pass version for clientSdk (#29)
extract chat-ui-react version to pass into `clientSdk` field in analytic event's payload by setting it through the internal function `addClientSdk` in ChatHeadless. This functionality is wrapped in a hook `useReportAnalyticsEvent`.

Due to the import of `package.json` the structure of `dist/` is reformatted (now all previous files are inside src/ and package.json is included in the respective esm and commonjs folder). As such, the entry points needs to be updated to reflect this.

J=CLIP-309
TEST=manual

see in test-site that clientSdk is passed into the payload with the correct versions (e.g. include the new CHAT_UI_REACT in addition to CHAT_HEADLESS_REACT, CHAT_HEADLESS and CHAT_CORE)

<img width="681" alt="Screenshot 2023-07-07 at 3 49 14 PM" src="https://github.com/yext/chat-ui-react/assets/36055303/821fa2d4-ddf5-4feb-9cb9-5b43720a8305">