ChatPopUp: support initial message pop up (#47)
support a new prop field in `ChatPopUp` call `showInitialMessagePopUp`. If true, `ChatPopUp` will display a popup next to the button showing the initial message. This message popup will NOT appear when once the panel is open (either on load or by manual click) or when there's an error fetching the initial message

Also refactored the fetch initial message into its own custom react hook for reusability

J=CLIP-743
TEST=manual&auto

see that tests passed
see that it works in test site

https://github.com/yext/chat-ui-react/assets/36055303/36e3e07a-e7af-47a1-8ba3-dcd9fed9cb68