GH workflows: linting, 3rd party license, and doc check (#58)
- Added workflow for: lint check, 3rd party license update and doc update.
- Added `BOT_REPO_SCOPED_TOKEN` repo secret for the bot to properly detect changes and auto-commit to repo as part of the GH workflows.

Downgraded generate-license-file from v2 to v1 to match the version in shared workflow (output format is slightly different). I tried to migrate the workflow to latest (v3) but noticed a bug so I submitted an issue report in their repo. Once that is done, the workflow version and the local devDep version of generate-license-file can be updated

J=CLIP-100
TEST=auto

see that the workflows ran as expected as checks under this PR