ChatUIReact: Chat bubble remembers user preference when reloading
The last open/close interaction made by the user is stored in local storage if the saveToLocalStorage flag is set in the config.
If this flag is set then the UI will either remain closed or opened depending on its status before the reload, even if there are messages.

J=CLIP-1310
TEST=manual

Deployed locally and unit tested